tests/requirements.txt
# Test framework dependencies
pyyaml>=6.0
pydantic>=2.0
typer>=0.9.0
rich>=13.0
